Externalize hardcoded config in loading, switch-audio, and credits-config for threlte-kit readiness #159
