+++ cspell.config.yaml
@@ -3,6 +3,7 @@ import:
   - "@joshuafolkken/kit/cspell/sveltekit"
 ignorePaths:
   - "**/credits-config.ts"
+  - "**/simon/credits.ts"
 words:
   - Baer
   - coeff
