+++ src/lib/game/FloorCredits.svelte
@@ -8,7 +8,6 @@
 		CREDITS_NORMAL_COLOR,
 		CREDITS_CYBER_COLOR,
 		CREDITS_POSITION_Y,
-		CREDITS_SCROLL_START_Z,
 		CREDITS_GLOW_BLUR,
 		CREDITS_GLOW_OPACITY,
 		FLOOR_TEXT_ROTATION_X
@@ -18,16 +17,18 @@
 	interface Props {
 		is_alt: boolean;
 		credits: string;
+		scroll_start_z: number;
+		scroll_end_z: number;
 	}
 
-	let { is_alt, credits }: Props = $props();
+	let { is_alt, credits, scroll_start_z, scroll_end_z }: Props = $props();
 
 	let current_font = $derived(fonts.get_font(is_alt));
 	let color = $derived(is_alt ? CREDITS_CYBER_COLOR : CREDITS_NORMAL_COLOR);
-	let scroll_z = $state(CREDITS_SCROLL_START_Z);
+	let scroll_z = $state(scroll_start_z);
 
 	function tick(delta: number): void {
-		scroll_z = advance_scroll(scroll_z, delta);
+		scroll_z = advance_scroll(scroll_z, delta, scroll_start_z, scroll_end_z);
 	}
 
 	useTask(tick);
