+++ src/lib/game/credits-config.ts
@@ -12,169 +12,8 @@ export const FLOOR_TEXT_ROTATION_X = -Math.PI / 2;
 
 const HALF_DIVISOR = 2;
 
-const CREDITS_LINES = [
-	'SIMON',
-	'CREDITS',
-	'',
-	'',
-	'',
-	'',
-	'SPONSORS',
-	'',
-	'',
-	'Dejisapo',
-	'',
-	'Incognito',
-	'',
-	'',
-	'',
-	'',
-	'GAME CONCEPT',
-	'',
-	'',
-	'Simon is inspired by the',
-	'Classic Electronic Memory Game',
-	'originally designed by',
-	'Ralph H. Baer',
-	'and Howard J. Morrison',
-	'published by Milton Bradley',
-	'in 1978',
-	'',
-	'',
-	'',
-	'',
-	'BUILT WITH',
-	'',
-	'',
-	'LANGUAGE',
-	'',
-	'TypeScript',
-	'',
-	'',
-	'FRAMEWORK',
-	'',
-	'Svelte 5 (Runes)',
-	'SvelteKit',
-	'',
-	'',
-	'3D RENDERING',
-	'',
-	'Three.js',
-	'Threlte',
-	'',
-	'',
-	'STYLING',
-	'',
-	'Tailwind CSS v4',
-	'',
-	'',
-	'BUILD TOOL',
-	'',
-	'Vite',
-	'',
-	'',
-	'AUDIO',
-	'',
-	'Web Audio API',
-	'',
-	'Sound Effect by',
-	'DRAGON-STUDIO',
-	'from Pixabay',
-	'',
-	'',
-	'DEPLOYMENT',
-	'',
-	'Cloudflare Workers',
-	'via Wrangler',
-	'',
-	'',
-	'',
-	'',
-	'FONTS',
-	'',
-	'',
-	'Press Start 2P',
-	'Cody "CodeMan38" Boisclair',
-	'SIL Open Font License',
-	'via Google Fonts',
-	'',
-	'Orbitron',
-	'Matt McInerney',
-	'SIL Open Font License',
-	'via Google Fonts',
-	'',
-	'',
-	'',
-	'',
-	'TESTING & TOOLING',
-	'',
-	'',
-	'Vitest',
-	'Playwright',
-	'ESLint',
-	'Prettier',
-	'pnpm',
-	'Lefthook',
-	'@joshuafolkken/kit',
-	'',
-	'',
-	'',
-	'',
-	'AI TOOLS',
-	'',
-	'',
-	'Claude Code',
-	'Sonnet 4.6',
-	'',
-	'',
-	'',
-	'',
-	'OPEN SOURCE LIBRARIES',
-	'',
-	'',
-	'svelte/svelte   MIT',
-	'sveltejs/kit   MIT',
-	'threlte/threlte   MIT',
-	'mrdoob/three.js   MIT',
-	'vitejs/vite   MIT',
-	'vitest-dev/vitest   MIT',
-	'microsoft/playwright   Apache 2.0',
-	'tailwindlabs/tailwindcss   MIT',
-	'',
-	'',
-	'',
-	'',
-	'SPECIAL THANKS',
-	'',
-	'',
-	'Unicorn',
-	'aopuyodaihuku',
-	'sakuramoti',
-	'@DocSeventeenth',
-	'@product_shokai',
-	'Longinus',
-	'fish',
-	'Ejirou',
-	'Satou',
-	'',
-	'',
-	'',
-	'',
-	'Built by Joshua Folkken',
-	'',
-	'',
-	'github.com/joshuafolkken/simon',
-	'',
-	'',
-	'',
-	'',
-	'',
-	'',
-	'THANK YOU FOR PLAYING !!'
-] as const;
-
-export const CREDITS_TEXT = CREDITS_LINES.join('\n');
-
-const CREDITS_TEXT_HEIGHT = CREDITS_LINES.length * CREDITS_FONT_SIZE * CREDITS_LINE_HEIGHT;
-export const CREDITS_SCROLL_START_Z = HALF_D + CREDITS_TEXT_HEIGHT / HALF_DIVISOR;
-export const CREDITS_SCROLL_END_Z = -(HALF_D + CREDITS_TEXT_HEIGHT / HALF_DIVISOR);
+export function make_credits_scroll_bounds(line_count: number): { start_z: number; end_z: number } {
+	const height = line_count * CREDITS_FONT_SIZE * CREDITS_LINE_HEIGHT;
+	const offset = HALF_D + height / HALF_DIVISOR;
+	return { start_z: offset, end_z: -offset };
+}
