+++ src/lib/simon/credits.ts
@@ -1,0 +1,168 @@
+import { make_credits_scroll_bounds } from '$lib/game/credits-config';
+
+const CREDITS_LINES = [
+	'SIMON',
+	'CREDITS',
+	'',
+	'',
+	'',
+	'',
+	'SPONSORS',
+	'',
+	'',
+	'Dejisapo',
+	'',
+	'Incognito',
+	'',
+	'',
+	'',
+	'',
+	'GAME CONCEPT',
+	'',
+	'',
+	'Simon is inspired by the',
+	'Classic Electronic Memory Game',
+	'originally designed by',
+	'Ralph H. Baer',
+	'and Howard J. Morrison',
+	'published by Milton Bradley',
+	'in 1978',
+	'',
+	'',
+	'',
+	'',
+	'BUILT WITH',
+	'',
+	'',
+	'LANGUAGE',
+	'',
+	'TypeScript',
+	'',
+	'',
+	'FRAMEWORK',
+	'',
+	'Svelte 5 (Runes)',
+	'SvelteKit',
+	'',
+	'',
+	'3D RENDERING',
+	'',
+	'Three.js',
+	'Threlte',
+	'',
+	'',
+	'STYLING',
+	'',
+	'Tailwind CSS v4',
+	'',
+	'',
+	'BUILD TOOL',
+	'',
+	'Vite',
+	'',
+	'',
+	'AUDIO',
+	'',
+	'Web Audio API',
+	'',
+	'Sound Effect by',
+	'DRAGON-STUDIO',
+	'from Pixabay',
+	'',
+	'',
+	'DEPLOYMENT',
+	'',
+	'Cloudflare Workers',
+	'via Wrangler',
+	'',
+	'',
+	'',
+	'',
+	'FONTS',
+	'',
+	'',
+	'Press Start 2P',
+	'Cody "CodeMan38" Boisclair',
+	'SIL Open Font License',
+	'via Google Fonts',
+	'',
+	'Orbitron',
+	'Matt McInerney',
+	'SIL Open Font License',
+	'via Google Fonts',
+	'',
+	'',
+	'',
+	'',
+	'TESTING & TOOLING',
+	'',
+	'',
+	'Vitest',
+	'Playwright',
+	'ESLint',
+	'Prettier',
+	'pnpm',
+	'Lefthook',
+	'@joshuafolkken/kit',
+	'',
+	'',
+	'',
+	'',
+	'AI TOOLS',
+	'',
+	'',
+	'Claude Code',
+	'Sonnet 4.6',
+	'',
+	'',
+	'',
+	'',
+	'OPEN SOURCE LIBRARIES',
+	'',
+	'',
+	'svelte/svelte   MIT',
+	'sveltejs/kit   MIT',
+	'threlte/threlte   MIT',
+	'mrdoob/three.js   MIT',
+	'vitejs/vite   MIT',
+	'vitest-dev/vitest   MIT',
+	'microsoft/playwright   Apache 2.0',
+	'tailwindlabs/tailwindcss   MIT',
+	'',
+	'',
+	'',
+	'',
+	'SPECIAL THANKS',
+	'',
+	'',
+	'Unicorn',
+	'aopuyodaihuku',
+	'sakuramoti',
+	'@DocSeventeenth',
+	'@product_shokai',
+	'Longinus',
+	'fish',
+	'Ejirou',
+	'Satou',
+	'',
+	'',
+	'',
+	'',
+	'Built by Joshua Folkken',
+	'',
+	'',
+	'github.com/joshuafolkken/simon',
+	'',
+	'',
+	'',
+	'',
+	'',
+	'',
+	'THANK YOU FOR PLAYING !!'
+] as const;
+
+export const CREDITS_TEXT = CREDITS_LINES.join('\n');
+
+const bounds = make_credits_scroll_bounds(CREDITS_LINES.length);
+export const CREDITS_SCROLL_START_Z = bounds.start_z;
+export const CREDITS_SCROLL_END_Z = bounds.end_z;
